Sistema Monetário e Player Update
Agora status do player em um script só
