+++ Projeto Cosmos/Assets/Prefabs/Asteroids Pack/Assets/Prefabs/Asteroid 1.prefab
@@ -274,7 +274,7 @@ Rigidbody:
   m_IsKinematic: 0
   m_Interpolate: 0
   m_Constraints: 126
-  m_CollisionDetection: 0
+  m_CollisionDetection: 1
 --- !u!114 &11425652
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -329,13 +329,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   maxHealth: 10
-  maxShield: 50
+  maxShield: 0
   health: 10
-  shield: 50
-  isOnCombatTimer: 5
+  shield: 0
   hasShield: 0
-  alive: 1
-  isOnCombat: 0
 --- !u!1 &176588
 GameObject:
   m_ObjectHideFlags: 0
