+++ Projeto Cosmos/Assets/Scenes/Portix.unity
@@ -909,7 +909,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
+      - m_Target: {fileID: 60776592}
         m_TargetAssemblyTypeName: PauseScript, Assembly-CSharp
         m_MethodName: Backtopause
         m_Mode: 1
@@ -1038,7 +1038,7 @@ RectTransform:
   - {fileID: 102862993}
   - {fileID: 2110956958}
   m_Father: {fileID: 0}
-  m_RootOrder: 13
+  m_RootOrder: 14
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -1845,7 +1845,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
+      - m_Target: {fileID: 60776592}
         m_TargetAssemblyTypeName: PauseScript, Assembly-CSharp
         m_MethodName: Resume
         m_Mode: 1
@@ -2434,7 +2434,7 @@ RectTransform:
   - {fileID: 1762009008}
   - {fileID: 624480985}
   m_Father: {fileID: 0}
-  m_RootOrder: 12
+  m_RootOrder: 13
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -3330,7 +3330,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
+      - m_Target: {fileID: 60776592}
         m_TargetAssemblyTypeName: PauseScript, Assembly-CSharp
         m_MethodName: Menu
         m_Mode: 1
@@ -4175,6 +4175,50 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 986890882}
   m_CullTransparentMesh: 1
+--- !u!1 &1077881783
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1077881785}
+  - component: {fileID: 1077881784}
+  m_Layer: 0
+  m_Name: MoneyManager
+  m_TagString: MoneyManager
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1077881784
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1077881783}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 82c4d7549a0e9a13080059b957bec68a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  money: 0
+--- !u!4 &1077881785
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1077881783}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: -10.455266, y: -5.7216845, z: -17.789639}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 12
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1084250348
 GameObject:
   m_ObjectHideFlags: 0
@@ -5000,28 +5044,6 @@ GameObject:
   m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 0be092eed109eed448ac411475f656e7, type: 3}
   m_PrefabInstance: {fileID: 207872821}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1383053155
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1383053154}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 26a7d24819554ae48983c72dfd00a324, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  healthBarScript: {fileID: 1619786452}
-  shieldBarScript: {fileID: 122782111}
-  maxHealth: 100
-  maxShield: 50
-  health: 150
-  shield: 50
-  isOnCombatTimer: 5
-  hasShield: 1
-  alive: 1
-  isOnCombat: 0
 --- !u!65 &1383053156
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -5123,6 +5145,28 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1.75, y: 0.50223565, z: 0.48896694}
   m_Center: {x: 0, y: 0.0625, z: -0.13051653}
+--- !u!114 &1383053165
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1383053154}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ea6dfc7e574f2e04688170631908402d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  healthBarScript: {fileID: 1619786452}
+  shieldBarScript: {fileID: 122782111}
+  maxHealth: 100
+  maxShield: 50
+  health: 100
+  shield: 50
+  isOnCombatTimer: 5
+  hasShield: 1
+  alive: 1
+  isOnCombat: 0
 --- !u!1 &1393777519
 GameObject:
   m_ObjectHideFlags: 0
@@ -6427,7 +6471,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
+      - m_Target: {fileID: 60776592}
         m_TargetAssemblyTypeName: PauseScript, Assembly-CSharp
         m_MethodName: Options
         m_Mode: 1
