+++ Projeto Cosmos/Assets/Scripts/PauseScript.cs
@@ -21,11 +21,13 @@ public class PauseScript : MonoBehaviour
         Cursor.lockState = CursorLockMode.Confined;
         PauseMenu.SetActive(false);
         OptionsMenu.SetActive(false);
+        Time.timeScale = 1f;
     }
     void Update()
     {
         if (Input.GetKeyDown(KeyCode.Escape) && !stationScript.isOnHUD)
         {
+            Debug.Log("teste");
             if (GamePaused)
             {
                 Resume();
@@ -51,6 +53,7 @@ public class PauseScript : MonoBehaviour
 
     void Pause()
     {
+        Debug.Log("teste Pause");
         PauseMenu.SetActive(true);
         OptionsMenu.SetActive(false);
         Time.timeScale = 0f;
