+++ Projeto Cosmos/Assets/Scripts/Portix/MoneyScript.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MoneyScript : MonoBehaviour
+{
+    public int money = 0;
+}
