+++ Projeto Cosmos/Assets/Scripts/Portix/MoneyScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 82c4d7549a0e9a13080059b957bec68a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
