+++ Projeto Cosmos/ProjectSettings/TagManager.asset
@@ -14,6 +14,7 @@ TagManager:
   - Asteroid
   - Drop
   - Missile
+  - MoneyManager
   layers:
   - Default
   - TransparentFX
